Make the build noarch again.
With the autogeneration of the pkg-config file, depending on the
architecture you build the package, the file gets installed in one of
the architecure dependent folders (/usr/lib64/pkgconfig,
/usr/lib/pkgconfig, /usr/lib/x86_64-linux-gnu/pkgconfig/, etc.). This
means it would require one build for architecture.

Make it noarch again (it's just headers) by forcing the installation of
the pkg-config file in /usr/share/pkgconfig.